requirements.txt
Flask
mysql-connector-python
openai
python-dotenv
